Refactor agency client
