+++ agency_client/src/agency_client.rs
@@ -3,7 +3,7 @@ use url::Url;
 
 use crate::agency_settings;
 use crate::error::{AgencyClientError, AgencyClientErrorKind, AgencyClientResult};
-use crate::mocking;
+use crate::testing::mocking;
 use crate::utils::{error_utils, validation};
 
 #[derive(Default, Deserialize)]
