+++ agency_client/src/utils/comm.rs
@@ -1,4 +1,5 @@
-use crate::{agency_settings, AgencyClientResult, httpclient};
+use crate::{agency_settings, AgencyClientResult};
+use crate::utils::httpclient;
 
 pub async fn post_to_agency(body_content: &Vec<u8>) -> AgencyClientResult<Vec<u8>> {
     let endpoint = agency_settings::get_config_value(agency_settings::CONFIG_AGENCY_ENDPOINT)?;
