--- agency_client/src/utils/timeout.rs
@@ -1,9 +1,0 @@
-use std::time::Duration;
-
-pub struct TimeoutUtils {}
-
-impl TimeoutUtils {
-    pub fn long_timeout() -> Duration {
-        Duration::from_secs(50)
-    }
-}
