+++ aries_vcx/src/utils/mod.rs
@@ -66,13 +66,13 @@ pub fn get_temp_dir_path(filename: &str) -> PathBuf {
 pub async fn send_message(sender_verkey: String, did_doc: DidDoc, message: A2AMessage) -> VcxResult<()> {
     trace!("send_message >>> message: {:?}, did_doc: {:?}", message, &did_doc);
     let envelope = EncryptionEnvelope::create(&message, Some(&sender_verkey), &did_doc).await?;
-    agency_client::httpclient::post_message(&envelope.0, &did_doc.get_endpoint()).await?;
+    agency_client::utils::httpclient::post_message(&envelope.0, &did_doc.get_endpoint()).await?;
     Ok(())
 }
 
 pub async fn send_message_anonymously(did_doc: &DidDoc, message: &A2AMessage) -> VcxResult<()> {
     trace!("send_message_anonymously >>> message: {:?}, did_doc: {:?}", message, &did_doc);
     let envelope = EncryptionEnvelope::create(&message, None, &did_doc).await?;
-    agency_client::httpclient::post_message(&envelope.0, &did_doc.get_endpoint()).await?;
+    agency_client::utils::httpclient::post_message(&envelope.0, &did_doc.get_endpoint()).await?;
     Ok(())
 }
